+++ tests/run/cpp_namespaces_helper.h
@@ -18,6 +18,11 @@ namespace A {
 
     typedef int A_t;
 
+    struct S {
+        A_t k;
+        double x;
+    };
+
     A_t A_func(A_t first, A_t second) {
         return first + second;
     }